Formatted the code
Looks much better now :D

+++ OsLab6/lab6release.c
@@ -1,5 +1,5 @@
 /**************************************************************************/
-/* PROGRAM: lab5child.c */
+/* PROGRAM: lab6release.c */
 /* DESCRIPTION: The child sleeps for a random time */
 /* and prints to the console once it terminates*/
 /**************************************************************************/
@@ -7,10 +7,10 @@
 //Includes
 # include <stdio.h>
 # include <stdlib.h>
-# include <sys/types.h>
-# include <sys/wait.h>
 # include <unistd.h>
-# include <time.h>   /* Needed for struct timespec */
+# include <sys/types.h>
+# include <fcntl.h>
+# include <errno.h>
 
 //Declare our helper functions
 void printUsage(char *programName);
@@ -23,13 +23,25 @@ int main(int argc, char *argv[])
     //First check the input
     checkInput(argc, argv);
 
-    char *fname; int fd, sleeptime, n_try, count=0; pid_t pid;
- pid = getpid(); srand((unsigned) pid); fname = argv[1]; sleeptime = atoi(argv[2]); n_try = atoi(argv[3]);
- while (unlink(fname)!=0)
- if (++count < n_try) sleep(sleeptime);
- else { printf ("\n Cannot release file\n"); exit(-1);
- }
- printf ("\n File is released\n"); 
+    //Initialize our variables
+    char *fname;
+    int fd, sleeptime, n_try, count=0;
+    pid_t pid;
+
+
+    pid = getpid();
+    srand((unsigned) pid);
+    fname = argv[1];
+    sleeptime = atoi(argv[2]);
+    n_try = atoi(argv[3]);
+
+
+    while (unlink(fname)!=0) if (++count < n_try) sleep(sleeptime);
+    else {
+        printf ("\n Cannot release file\n"); exit(-1);
+    }
+
+    printf ("\n File is released\n");
 
 }
